Move grammar loading to a dedicated crate
This will allow building tools that work on the grammar.
